move not-found to errors folder
